% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scclusterboot.R
\name{AssignHighestJaccard}
\alias{AssignHighestJaccard}
\title{Assign highest Jaccard index for each cluster of the subsampled data set before
reclustering with the cluster identites of subsampled data set after reclustering}
\usage{
AssignHighestJaccard(idents1, idents2)
}
\arguments{
\item{idents1}{A list of cluster identity copied from the orginal data sets.
idents1 is a list of the cluster identity from the subsampled data sets before reclustering.}

\item{idents2}{A list of cluster identity from the subsampled data sets.
idents2 is a list of the cluster identity from the subsampled data sets after reclustering.
The order of identities in idents1 and idents2 should correspond to each other.}
}
\value{
A matrix with dimention of #number of subsampling * #number of clusters in the
original data set.
}
\description{
Assign highest Jaccard index for each cluster of the subsampled data set before
reclustering with the cluster identites of subsampled data set after reclustering
}
