% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterviz.R
\name{BootParameterScatterPlot}
\alias{BootParameterScatterPlot}
\title{Scatter plot of number of stable clusters across different k
and percentage of cells in stable clusters across different k.}
\usage{
BootParameterScatterPlot(stable_cluster)
}
\arguments{
\item{stable_cluster}{A list of list returned by AssignStableCluster}
}
\value{
A ggplot object
}
\description{
Scatter plot of number of stable clusters across different k
and percentage of cells in stable clusters across different k.
}
\examples{
\dontrun{
## see README.md
AssignStableCluster(ks_idents_original$k20, ks_idents$`20`)

ks_stable<- purrr::map2(ks_idents_original, ks_idents, ~AssignStableCluster(ident1= .x, idents = .y))
}
}
