% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scclusterboot.R
\name{CalculatePercentCellInStable}
\alias{CalculatePercentCellInStable}
\title{Calculate the percentage of cells in stable clusters in the full data set}
\usage{
CalculatePercentCellInStable(ident, stable_cluster)
}
\arguments{
\item{ident.}{A named factor vector. names are the cell names, the values are
the cluster id from the full data set.}

\item{stable_cluster.}{A logical vector for each of the original cluster indicating
it is stable or not, calculated from \code{\link{AssignStableCluster}}}
}
\value{
A percentage of cells in stable cluster
}
\description{
Calculate the percentage of cells in stable clusters in the full data set
}
