% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatesilhouette.R
\name{CalculateSilhouette}
\alias{CalculateSilhouette}
\title{Calculate Silhouette width from PCA space for each cell after clustering
This is calculated from Seurat object}
\usage{
CalculateSilhouette(object, dims = 1:50)
}
\arguments{
\item{object}{A Seurat object with Idents set to cluster ids (factors)}

\item{dims}{default 1:50  dimension to use in the PCA space to calculate
eucledian distance}
}
\value{
a dataframe with silhouette width for each cell. see also \code{\link[cluster]{silhouette}}
}
\description{
Calculate Silhouette width from PCA space for each cell after clustering
This is calculated from Seurat object
}
\examples{
CalculateSilhouette(pbmc_small, dims = 1:15)

}
