% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterviz.R
\name{ClusterSizeBarplot}
\alias{ClusterSizeBarplot}
\title{Make a Barplot for cluster size}
\usage{
ClusterSizeBarplot(ident, bar_col = "blue")
}
\arguments{
\item{ident}{a named factor vector. names are the cell names, the values are
the cluster id.}

\item{bar_col}{color for the bar. Default is blue.}
}
\value{
a ggplot2 bar graph object
}
\description{
Make a Barplot for cluster size
}
\examples{
data(pbmc_small)
CLusterSizeBarplot(pbmc_small@ident)

}
