% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomflatviolin.R
\name{geom_flat_violin}
\alias{geom_flat_violin}
\title{A Flat Violin plot}
\usage{
geom_flat_violin(mapping = NULL, data = NULL, stat = "ydensity",
  position = "dodge", trim = TRUE, scale = "area",
  show.legend = NA, inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{See \code{\link[ggplot2]{geom_violin}}}

\item{data}{See \code{\link[ggplot2]{geom_violin}}}

\item{position}{See \code{\link[ggplot2]{geom_violin}}}

\item{trim}{See \code{\link[ggplot2]{geom_violin}}}

\item{scale}{See \code{\link[ggplot2]{geom_violin}}}

\item{show.legend}{See \code{\link[ggplot2]{geom_violin}}}

\item{inherit.aes}{See \code{\link[ggplot2]{geom_violin}}}

\item{...}{}
}
\description{
A Flat Violin plot
}
\examples{

ggplot(diamonds, aes(cut, carat)) +
geom_flat_violin() +
coord_flip()

}
