% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterviz.R
\name{JaccardRainCloudPlot}
\alias{JaccardRainCloudPlot}
\title{Plot the Jaccard index distribution using raincloud plot}
\usage{
JaccardRainCloudPlot(idents1, idents2, title = NULL)
}
\arguments{
\item{idents1}{A list of cluster identity from the subsampled data set
before reclustering. (cluster id copied from the original full data set)}

\item{idents2}{A list of cluster identity from the subsampled data sets after
reclustering.}

\item{title}{Title of the plot}
}
\value{
A ggplot2 object
}
\description{
Plot the Jaccard index distribution using raincloud plot
}
\examples{

\dontrun{
data(idents)
## the pbmc here need to be fully processed.
JaccardRainCloudPlot(idents, idents)
}

}
