% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scclusterboot.R
\name{JaccardSets}
\alias{JaccardSets}
\title{Calculate jaccard distance for two sets of character vectors}
\usage{
JaccardSets(set1, set2)
}
\arguments{
\item{set1}{character vector 1}

\item{set2}{character vector 2}
}
\value{
jaccard distance
}
\description{
Calculate jaccard distance for two sets of character vectors
}
\examples{
JaccardSets(sample(LETTERS, 10), sample(LETTERS, 10))
}
