% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scclusterboot.R
\name{MatchClusters}
\alias{MatchClusters}
\title{Match two run of cluster ids with highest Jaccard index}
\usage{
MatchClusters(ident1, ident2)
}
\arguments{
\item{ident1}{a named factor vector. names are the cell names, the values are
the cluster id.}

\item{ident2}{a named factor vector. names are the cell names, the values are
the cluster id.}
}
\value{
A tibble with two columns, column 1 is the cluster ids from ident1, column2
is the cluster ids from ident2.
}
\description{
Match two run of cluster ids with highest Jaccard index
}
\examples{
 MatchClusters(pbmc@ident, pbmc_small@ident)
}
