% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergemultipleseuratobjects.R
\name{MergeMultipleSeuratObjects}
\alias{MergeMultipleSeuratObjects}
\title{Read multiple 10x run into Seurat objects and merge into a single Seurat object}
\usage{
MergeMultipleSeuratObjects(input_folders, do.normalize = FALSE, ...)
}
\arguments{
\item{input_folders}{A named list of folder path for each run.}

\item{do.normalize}{Whether or not normalize the data after mergeing, default is FALSE}

\item{...}{Other parameters for CreatSeuratObject in the Seurat package}
}
\value{
A single merged Seurat object from mulitple 10x runs.
}
\description{
Read multiple 10x run into Seurat objects and merge into a single Seurat object.
The names of the list of paths will be prepended to the cell name.
}
\examples{
\dontrun{
library(fs)
library(here)
library(stringr)
input_folders<- dir_ls( path = here("data"), recursive = T) \%>\% path_dir() \%>\%
unique() \%>\% str_subset("mm10-1.2.0_premrna")
merged_seurat<- MergeMultipleSeuratObjects(input_folders)
}
}
