% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scclusterboot.R
\name{PairWiseJaccardSets}
\alias{PairWiseJaccardSets}
\title{Calculate pair-wise Jaccard distance for @ident slots from two Seurat objects}
\usage{
PairWiseJaccardSets(ident1, ident2)
}
\arguments{
\item{ident1}{a named factor vector. names are the cell names, the values are
the cluster id.}

\item{ident2}{a named factor vector. names are the cell names, the values are
the cluster id.}
}
\value{
a matrix of pair-wise Jaccard distance. Rows are clusters from ident1,
columns are clusters from ident2
}
\description{
Calculate pair-wise Jaccard distance for two named factor vector. e.g.
seurat_obj1@ident and seurat_obj2@ident
}
\examples{
PairWiseJaccardSets(pbmc@ident, pbmc_small@ident)


}
