% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterviz.R
\name{PairWiseJaccardSetsHeatmap}
\alias{PairWiseJaccardSetsHeatmap}
\title{Make a Heatmap of the pairwise Jaccard distance between cluster ident of two
Seurat object}
\usage{
PairWiseJaccardSetsHeatmap(ident1, ident2, title = NULL,
  col_low = "white", col_high = "red", cluster_rows = F,
  cluster_columns = F, show_row_dend = T, show_column_dend = T)
}
\arguments{
\item{ident1}{a named factor vector. names are the cell names, the values are
the cluster id.}

\item{ident2}{a named factor vector. names are the cell names, the values are
the cluster id.}

\item{title}{The title of the heatmap}

\item{col_low}{Color for low Jaccard index.}

\item{col_high}{Color for high Jaccard index.}

\item{cluster_rows}{cluster row or not, default FALSE}

\item{cluster_columns}{cluster columns or not, default FASLE}

\item{show_row_dend}{Whether or not show row dendrogram}

\item{show_column_dend}{Whether or not show column dendrogram}
}
\value{
A Heatmap representing the pair-wise Jaccard correlation
}
\description{
Make a Heatmap of the pairwise Jaccard distance between cluster ident of two
Seurat object
}
\examples{

}
