% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scclusterboot.R
\name{PairWiseOverlappingIdents}
\alias{PairWiseOverlappingIdents}
\title{Calculate pair-wise overlapping cluster identities for @ident slots from two Seurat objects}
\usage{
PairWiseOverlappingIdents(ident1, ident2)
}
\arguments{
\item{ident1}{a named factor vector. names are the cell names, the values are
the cluster id.}

\item{ident2}{a named factor vector. names are the cell names, the values are
the cluster id.}
}
\value{
A matrix of pairwise number of common cell identities for each cluster.
}
\description{
Calculate pair-wise overlapping cluster identities for two named factor vector. e.g.
seurat_obj1@ident and seurat_obj2@ident
}
\examples{
PairWiseOverlappingIdents(pbmc@ident, pbmc_small@ident)
}
