% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterviz.R
\name{ParameterSetScatterPlot}
\alias{ParameterSetScatterPlot}
\title{Plot a scatter plot for different clustering parameters}
\usage{
ParameterSetScatterPlot(
  stable_clusters,
  fullsample_idents,
  x_var,
  y_var,
  facet_rows,
  facet_cols
)
}
\arguments{
\item{stable_clusters}{a dataframe with list-columns for data, stable_cluster determined by
\code{\link{AssignStableCluster}} and the rest of the columns are pc, resolution and k_param.}

\item{fullsample_idents}{a dataframe with the list-column contain the original ident for
the full dataset. This is the direct output from the Snakemake workflow.}

\item{x_var}{one of "pc", "resolution" and "k_param".}

\item{y_var}{one of "number" or "percentage". If it is "number",
y-axis si the total number of clusters and total number of stable clusters.}

\item{facet_rows}{one of "pc", "resolution" and "k_param" for ggplot2 to facet.}

\item{facet_cols}{one of "pc", "resolution" and "k_param" for ggplot2 to facet.}
}
\value{
a ggplot2 object
}
\description{
x-axis is the parameters tested (e.g. many different k.param)
y-axis is the total number of clusters and total number of stable clusters based
on the jaccard cutoff as determined by AssignStableClusters, or precentage of cells
in stable clusters.
}
