% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocesssubsetdata.R
\name{PreprocessSubsetData}
\alias{PreprocessSubsetData}
\title{A wrapper for preprocessing subsetted Seurat object}
\usage{
PreprocessSubsetData(object, mean.function = ExpMean,
  dispersion.function = logVMR, x.low.cutoff = 0.05,
  x.high.cutoff = 10, y.cutoff = 0.5, ...)
}
\arguments{
\item{object}{A subsetted Seurat object created by RandomSubsetData}

\item{mean.function}{Function to compute x-axis value (average expression).
Default is to take the mean of the detected (i.e. non-zero) values}

\item{dispersion.function}{Function to compute y-axis value (dispersion).
Default is to take the standard deviation of all values}

\item{x.low.cutoff}{Bottom cutoff on x-axis for identifying variable genes}

\item{x.high.cutoff}{Top cutoff on x-axis for identifying variable genes}

\item{y.cutoff}{Bottom cutoff on y-axis for identifying variable genes}

\item{...}{any other parameters for FindVariableGenes}
}
\value{
a fully processed Seurat object
}
\description{
The wrapper does FindVeriableGenes, ScaleData, RunPCA, JackStraw to
determine how many PCs to use, ProjectPCA and FindClusters and retrun
a fully processed Seurat object. The input subsetted seurat object is
supposed to be fully processed as well. So the NormalizeData step is not
necessary.
}
\examples{
pbmc_small_subset<- RandomSubsetData(pbmc_small, 0.8)
pbmc_small_subset_processed<- PreprocessSubsetData(pbmc_small_subset)
pbmc_small_subset_processed@meta.data
}
