% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocesssubsetdatav2.R
\name{PreprocessSubsetDataV2}
\alias{PreprocessSubsetDataV2}
\title{A wrapper for preprocessing subsetted Seurat object using ScaleData}
\usage{
PreprocessSubsetDataV2(object, nfeatures = 2000, num.pc = 20,
  pc.use = NULL, workers = 2, score.thresh = 1e-05,
  sig.pc.thresh = 0.05, n.start = 100, nn.eps = 0,
  resolution = 0.8, k.param = 30, ...)
}
\arguments{
\item{object}{A subsetted Seurat object created by RandomSubsetData}

\item{num.pc}{number of PCs to calculate in RunPCA, JackStraw and JackStrawPlot
step. The optimal PCs for FindClusters will be determined by only significant PCs
from JackStrawPlot or if pc.use is set, JackStraw step will be skipped and use pc.use
for FindClusters.}

\item{pc.use}{number of PCs used for FindClusters. if pc.use is set, JackStraw step
will be skipped and use pc.use for FindClusters. score.thresh and sig.pc.thresh will be ignored.}

\item{workers}{number of CPUs to use for \code{\link[future]{plan}} parallel processing}

\item{score.thresh}{Threshold to use for the proportion test of PC significance.}

\item{sig.pc.thresh}{Threshold for the significance of a particular PC.}

\item{n.start}{Number of random start.}

\item{nn.eps}{Error bound when performing nearest neighbor seach using RANN;
default of 0.0 implies exact nearest neighbor search. See FindClusters.}

\item{resolution}{Value of the resolution parameter, use a value above (below)
1.0 if you want to obtain a larger (smaller) number of communities. see FIndClusters.}

\item{k.param}{Defines k for the k-nearest neighbor algorithm.}

\item{...}{any other parameters}

\item{variable.features.n}{number of variable features for \code{\link[Seurat]{SCTransform}}}
}
\value{
a fully processed Seurat object
}
\description{
The wrapper does FindVeriableGenes, ScaleData, RunPCA, JackStraw to
determine how many PCs to use, ProjectPCA and FindClusters and retrun
a fully processed Seurat object. The input subsetted seurat object is
supposed to be fully processed as well. So the NormalizeData step is not
necessary.
}
\examples{
pbmc_small_subset<- RandomSubsetData(pbmc_small, 0.8)
pbmc_small_subset_processed<- PreprocessSubsetData(pbmc_small_subset)
pbmc_small_subset_processed@meta.data
}
