% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomsubsetdata.R
\name{RandomSubsetData}
\alias{RandomSubsetData}
\title{Randomly subset (cells) seurat object by a rate}
\usage{
RandomSubsetData(object, rate, random.seed = 1, do.clean = TRUE, ...)
}
\arguments{
\item{object}{Seurat object}

\item{rate}{a number betwee 0-1 for subsetting}

\item{random.seed}{set a random seed for sampling, default is 1.}

\item{do.clean}{subset the seurat object in a clean way, include the raw.data}

\item{...}{any other parameters to SubsetData from Seurat package}
}
\value{
Returns a randomly subsetted seurat object
}
\description{
Randomly subset (cells) seurat object by a rate
}
\examples{
pbmc_small
pbmc_small_subset<- RandomSubsetData(pbmc_small, 0.8)
dim(pbmc_small_subset@meta.data)

}
