% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scclusterboot.R
\name{scClusterBoot}
\alias{scClusterBoot}
\title{Bootstrap for a fully processed Seurat object}
\usage{
scClusterBoot(object, n = 4, rate = 0.8, ...)
}
\arguments{
\item{object}{A fully processed Seurat object.}

\item{n}{Number of times you want to bootstrap.}

\item{rate}{A number between 0 and 1 for subsampling the cells.}

\item{...}{Other parameters passed to \code{\link{PreprocessSubsetData}}}
}
\value{
A list of lists containing the ident from the subsetted reclustered
seurat objects.
}
\description{
Bootstrap for a fully processed Seurat object
}
\examples{

}
