% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scclusterboot.R
\name{SelectHighestJaccard}
\alias{SelectHighestJaccard}
\title{Select Highest Jaccard to match clusters}
\usage{
SelectHighestJaccard(mat)
}
\arguments{
\item{mat}{This matrix is returned by \code{\link{PairWiseJaccardSets}}}
}
\value{
A named number of Jaccard distance with names of the original cluster id
in the original unsubsetted seurat object.
}
\description{
Select Highest Jaccard to match clusters
}
