% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterviz.R
\name{SilhouetteRainCloudPlot}
\alias{SilhouetteRainCloudPlot}
\title{Plot raincloud plot for silhouette score}
\usage{
SilhouetteRainCloudPlot(silhouette_score)
}
\arguments{
\item{silhouette_score}{a dataframe returned by \code{link[CalculateSilhouette]}}
}
\value{
a ggplot2 object
}
\description{
Plot raincloud plot for silhouette score
}
\examples{

SilhouetteRainCloudPlot(CalculateSilhouette(pbmc_small, dims = 1:15))
}
