% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scclusterboot.R
\name{AssignStableCluster}
\alias{AssignStableCluster}
\title{Assign stable cluster}
\usage{
AssignStableCluster(ident1, idents, method = median, cutoff = 0.4)
}
\arguments{
\item{ident1}{The cluster identity from the original full data set.}

\item{idents}{A list of cluster identity from the subsampled data sets.}

\item{method}{what function to summarize the jaccard index across all simulations.
default is median}

\item{cutoff}{Cutoff for assigning stable clusters. A jaccard of 0.4 is used
for default}
}
\value{
A list containing the raw data for jaccard index for all simulations,
TRUE or FALSE of stable cluster for each cluster and a number of stable clusters
and percentage of cells in stable clusters.
}
\description{
Assign stable cluster
}
\examples{

data(idents)
AssignStableCluster(pbmc@ident, idents)
}
