% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scclusterboot.R
\name{AssignStableCluster}
\alias{AssignStableCluster}
\title{Assign stable cluster}
\usage{
AssignStableCluster(idents1, idents2, method = median, cutoff = 0.4)
}
\arguments{
\item{idents1}{A list of cluster identity copied from the orginal data sets.
idents1 is a list of the cluster identity from the subsampled data sets before reclustering.}

\item{idents2}{A list of cluster identity from the subsampled data sets.
idents2 is a list of the cluster identity from the subsampled data sets after reclustering.
The order of identities in idents1 and idents2 should correspond to each other.}

\item{method}{what function to summarize the jaccard index across all simulations.
default is median}

\item{cutoff}{Cutoff for assigning stable clusters. A jaccard of 0.4 is used
for default}
}
\value{
A list containing the raw data for jaccard index for all simulations,
TRUE or FALSE of stable cluster for each cluster and a number of stable clusters.
}
\description{
Assign stable cluster
}
\examples{

data(idents)
dummy example, all clusters are stable.
AssignStableCluster(idents, idents)

}
