% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterviz.R
\name{BootParameterScatterPlot}
\alias{BootParameterScatterPlot}
\title{Scatter plot of number of stable clusters across different k
and percentage of cells in stable clusters across different k.}
\usage{
BootParameterScatterPlot(ks_stable, ks_idents_original)
}
\arguments{
\item{ks_stable}{A list of list returned by AssignStableCluster}

\item{ks_idents_original}{A list of idents of the full data sets for each k}
}
\value{
A ggplot object
}
\description{
Scatter plot of number of stable clusters across different k
and percentage of cells in stable clusters across different k.
}
\examples{
\dontrun{
## see README.md
AssignStableCluster(ks_idents_original$k20, ks_idents$`20`)

ks_stable<- purrr::map2(ks_idents_original, ks_idents, ~AssignStableCluster(ident1= .x, idents = .y))
}
}
