% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterviz.R
\name{JaccardRainCloudPlot}
\alias{JaccardRainCloudPlot}
\title{Plot the Jaccard index distribution using raincloud plot}
\usage{
JaccardRainCloudPlot(ident1, idents, title = NULL)
}
\arguments{
\item{ident1}{The cluster identity from the original full data set.}

\item{idents}{A list of cluster identity from the subsampled data sets.}

\item{title}{Title of the plot}
}
\value{
A ggplot2 object
}
\description{
Plot the Jaccard index distribution using raincloud plot
}
\examples{

\dontrun{
data(idents)
## the pbmc here need to be fully processed.
JaccardRainCloudPlot(pbmc@ident, idents)
}

}
