% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterviz.R
\name{PairWiseJaccardSetsHeatmap}
\alias{PairWiseJaccardSetsHeatmap}
\title{Make a Heatmap of the pairwise Jaccard distance between cluster ident of two
Seurat object}
\usage{
PairWiseJaccardSetsHeatmap(mat, col_low = "white", col_high = "red")
}
\arguments{
\item{mat}{The pairwise Jaccard distance matrix returned by
\code{\link{PairWiseJaccardSets}}}

\item{col_low}{Color for low Jaccard index.}

\item{col_high}{Color for high Jaccard index.}
}
\value{
A Heatmap representing the pair-wise Jaccard correlation
}
\description{
Make a Heatmap of the pairwise Jaccard distance between cluster ident of two
Seurat object
}
\examples{

}
