% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomsubsetdata.R
\name{RandomSubsetData}
\alias{RandomSubsetData}
\title{Randomly subset (cells) seurat object by a rate}
\usage{
RandomSubsetData(object, rate, random.subset.seed = NULL, ...)
}
\arguments{
\item{object}{Seurat object}

\item{rate}{a number betwee 0-1 for subsetting}

\item{random.subset.seed}{set a random seed for sampling, default is NULL.}

\item{...}{any other parameters to \code{\link[Seurat]{subset}}}
}
\value{
Returns a randomly subsetted seurat object
}
\description{
Randomly subset (cells) seurat object by a rate
}
\examples{
pbmc_small
pbmc_small_subset<- RandomSubsetData(pbmc_small, 0.8)
dim(pbmc_small_subset@meta.data)


}
