% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scclusteval-package.R
\docType{package}
\name{scclusteval-package}
\alias{scclusteval}
\alias{scclusteval-package}
\title{scclusteval: Evaluate the single cell clustering}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/crazyhottommy/scclusteval}
  \item Report bugs at \url{https://github.com/crazyhottommy/scclusteval/issues}
}

}
\author{
\strong{Maintainer}: Ming Tang \email{tangming2005@gmail.com}

}
\keyword{internal}
